chore: move timeoutContent
