+++ apps/barry/src/modules/developers/commands/chatinput/devtools/tools/unblacklist.ts
@@ -2,7 +2,7 @@ import type { UpdatableInteraction } from "@barry/core";
 import type DevelopersModule from "../../../../index.js";
 
 import { ComponentType, TextInputStyle } from "@discordjs/core";
-import { timeoutContent } from "../../../../../marketplace/constants.js";
+import { timeoutContent } from "../../../../../../common.js";
 import config from "../../../../../../config.js";
 
 /**
