+++ apps/barry/src/modules/marketplace/dependencies/requests/editor/RequestEditor.ts
@@ -4,10 +4,10 @@ import type { UpdatableInteraction } from "@barry/core";
 import type RequestsModule from "../index.js";
 
 import { ButtonStyle, ComponentType, MessageFlags } from "@discordjs/core";
-import { INVITE_REGEX, retryComponents, timeoutContent } from "../../../constants.js";
+import { INVITE_REGEX, retryComponents } from "../../../constants.js";
 import { capitalizeEachSentence, capitalizeEachWord, displayContact } from "../../../utils.js";
 import { getEditContactContent, getEditRequestContent, getRequestContent } from "./functions/content.js";
-
+import { timeoutContent } from "../../../../../common.js";
 import config from "../../../../../config.js";
 
 /**
