+++ apps/barry/tests/modules/marketplace/requests/events/editRequest.test.ts
@@ -11,7 +11,7 @@ import {
 import { RequestEditor } from "../../../../../src/modules/marketplace/dependencies/requests/editor/RequestEditor.js";
 import { createMockApplication } from "../../../../mocks/index.js";
 import { mockRequest } from "../mocks/request.js";
-import { timeoutContent } from "../../../../../src/modules/marketplace/constants.js";
+import { timeoutContent } from "../../../../../src/common.js";
 
 import RequestsModule, { ManageRequestButton } from "../../../../../src/modules/marketplace/dependencies/requests/index.js";
 import EditRequestEvent from "../../../../../src/modules/marketplace/dependencies/requests/events/editRequest.js";
